AddAdminsRole

+++ MiniMarket/Models/OrderItem.cs
@@ -11,5 +11,5 @@ public class OrderItem
     public int ProductId { get; set; }
     public int ProductPrice { get; set; }
     public int Quantity { get; set; }
-    public decimal UnitPrice { get; set; }
+    public int UnitPrice { get; set; }
 }
